+++ CHANGELOG.md
@@ -21,6 +21,7 @@ hangeLog
 - BUGFIX: fix "t2t-sentence" node by finally using the "timeout" parameter at all
 - BUGFIX: fix "t2t-sentence" node by correctly adjusting timestamp end on merged chunks
 - BUGFIX: fix "t2t-sentence" node by doing a more precide decision for preview chunks
+- BUGFIX: fix "t2t-sentence" node by fixing Queue class to adjust foreign affected pointers on insert/delete
 - CLEANUP: cleanup result type of audio classification pipeline in gender node
 
 2.3.0 (2026-03-27)